arreglos y funcionalidad total

+++ src/models/domicilios.go
@@ -4,7 +4,7 @@ type Domicilios  struct {
 	IDOrdenGeneral int    `gorm:"column:idOrdenGeneral"`
 	NombreCliente  string `gorm:"column:nombreCliente"`
 	DireccionOrden string `gorm:"column:direccionOrden"`
-	TotalOrden     int    `gorm:"column:totalOrden"`
+	TotalOrden     float64    `gorm:"column:totalOrden"`
 	FechaCrea      string `gorm:"column:fechaCrea"`
 	NombreTraza    string `gorm:"column:nombreTraza"`
 	NombreDomiciliario string `gorm:"column:domiciliario"`
